update auth controller
