+++ back/.env.example
@@ -13,3 +13,5 @@ GITHUB_CLIENT_ID=
 GITHUB_CLIENT_SECRET=
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
+BACK_URL=http://localhost:3333
+FRONT_URL=http://localhost:5173
