+++ back/config/ally.ts
@@ -5,17 +5,17 @@ const allyConfig = defineConfig({
   github: services.github({
     clientId: env.get('GITHUB_CLIENT_ID'),
     clientSecret: env.get('GITHUB_CLIENT_SECRET'),
-    callbackUrl: 'http://localhost:3333/github/callback',
+    callbackUrl: `${env.get('BACK_URL')}/github/callback`,
   }),
   google: services.google({
     clientId: env.get('GOOGLE_CLIENT_ID'),
     clientSecret: env.get('GOOGLE_CLIENT_SECRET'),
-    callbackUrl: '',
+    callbackUrl: `${env.get('BACK_URL')}/google/callback`,
   }),
 })
 
 export default allyConfig
 
 declare module '@adonisjs/ally/types' {
-  interface SocialProviders extends InferSocialProviders<typeof allyConfig> { }
+  interface SocialProviders extends InferSocialProviders<typeof allyConfig> {}
 }
