+++ back/start/routes.ts
@@ -9,7 +9,6 @@
 
 import router from '@adonisjs/core/services/router'
 import { middleware } from './kernel.js'
-import User from '#models/user'
 
 const AuthController = () => import('#controllers/auth_controller')
 
@@ -45,58 +44,4 @@ router
   })
   .where('provider', /github|google/)
 
-router
-  .get('/:provider/callback', async ({ ally, params, response }) => {
-    const driverInstance = ally.use(params.provider)
-
-    /**
-     * User has denied access by canceling
-     * the login flow
-     */
-    if (driverInstance.accessDenied()) {
-      return 'You have cancelled the login process'
-    }
-
-    /**
-     * OAuth state verification failed. This happens when the
-     * CSRF cookie gets expired.
-     */
-    if (driverInstance.stateMisMatch()) {
-      return 'We are unable to verify the request. Please try again'
-    }
-
-    /**
-     * GitHub responded with some error
-     */
-    if (driverInstance.hasError()) {
-      return driverInstance.getError()
-    }
-
-    /**
-     * Access user info
-     */
-    const user = await driverInstance.user()
-
-    let dbUser = await User.findBy('email', user.email)
-    if (!dbUser) {
-      dbUser = await User.create({
-        email: user.email,
-        username: user.name,
-      })
-    }
-
-    const accessToken = await User.accessTokens.create(dbUser)
-    const token = accessToken.toJSON().token
-
-    response.plainCookie('session', token, {
-      httpOnly: true,
-      secure: true,
-      sameSite: 'lax',
-      path: '/',
-      maxAge: 60 * 60 * 24 * 30,
-      encode: false,
-    })
-
-    return response.redirect('http://localhost:5173/')
-  })
-  .where('provider', /github|google/)
+router.get('/:provider/callback', [AuthController, 'callback']).where('provider', /github|google/)
